whatsapp-bot/requirements.txt
blis==0.7.4
cachetools==4.2.1
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
dialogflow==1.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
Flask==1.1.2
google-api-core==1.25.1
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-translate==2.0.1
googleapis-common-protos==1.52.0
grpcio==1.35.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
MarkupSafe==1.1.1
murmurhash==1.0.5
numpy==1.19.0
plac==1.1.3
preshed==3.0.5
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2020.5
regex==2020.11.13
requests==2.25.1
rsa==4.7
Shapely==1.7.1
six==1.15.0
soton-corenlppy==1.0.0
spacy==2.3.5
srsly==1.0.5
thinc==7.4.5
tqdm==4.56.0
twilio==6.51.0
urllib3==1.26.3
uWSGI==2.0.19.1
wasabi==0.8.2
Werkzeug==1.0.1
